+++ src/hkimjp/konpy2/admin.clj
@@ -1,13 +1,26 @@
 (ns hkimjp.konpy2.admin
   (:require
+   [java-time.api :as jt]
+   [hiccup2.core :as h]
+   [ring.util.anti-forgery :refer [anti-forgery-field]]
+   [ring.util.response :as resp]
    [taoensso.telemere :as t]
    [hkimjp.datascript :as ds]
-   [hkimjp.konpy2.response :refer [page]]))
+   [hkimjp.konpy2.response :refer [page]]
+   [hkimjp.konpy2.util :refer [user]]))
+
+(def btn "p-1 rounded text-white bg-sky-500 hover:bg-sky-700 active:bg-red-500")
+
+(def btn-admin "p-1 rounded-xl text-white bg-red-500 hover:bg-red-700 active:bg-red-900")
+
+(def box "text-center size-10 outline  shadow-lg outline-black/5")
+
+(def te "my-2 p-2 text-md font-mono grow h-60 outline outline-black")
 
 (defn admin [_request]
   (page
    [:div
-    [:div "admin only"]
+    [:div.text-2xl.font-bold "admin only"]
     [:ul
      [:li [:a {:href "/admin/problems"} "problems"]]
      [:li [:a {:href "/admin/new"} "new"]]
@@ -27,33 +40,36 @@
 
 (defn upsert [])
 
-(defn create! [{params :params}]
-  (t/log! :info "create! params")
-  (page
-   [:div "create!"]))
+(defn- div-textarea [label text]
+  [:textarea.w-full.h-20.p-2.outline.outline-black.shadow-lg
+   {:name label} text])
 
-(def ta-class "w-180 h-10 p-2 outline outline-black/5 shadow-lg")
+(defn- section [title]
+  [:div.font-bold.pu-4 title])
+
+(defn- input-box [label val]
+  [:input.text-center.size-6.outline {:name label :value val}])
 
 (defn- problem-form
-  [{:keys [db/id
-           problem/valid
-           week
-           num
-           problem
-           test
-           gpt] :as params}]
-  [:form {:method "post" :action "/admin/create"}
-   [:input {:type "hidden" :name "db/id" :value id}]
-   [:input {:type "hidden" :name "problem/valid" :value valid}]
-   [:div
-    [:input {:name "week" :value week}] "-" [:input {:name "num" :value num}]]
-   [:div [:textarea {:class ta-class :name "problem"} problem]]
-   [:div [:textarea {:class ta-class :name "test"} test]]
-   [:div [:textarea {:class ta-class :name "gpt"} gpt]]
-   [:button "create"]])
+  [{:keys [db/id problem/valid week num problem test gpt] :as params}]
+  [:div
+   [:div.text-2xl.font-bold "Problem-Form"]
+   [:form.mx-4 {:method "post"}
+    (h/raw (anti-forgery-field))
+    [:input {:type "hidden" :name "db/id" :value id}]
+    [:input {:type "hidden" :name "problem/valid" :value valid}]
+    (section "week-num")
+    [:div (input-box "week" week) " - " (input-box "num" num)]
+    (section "problem")
+    (div-textarea "problem" problem)
+    (section "test")
+    (div-textarea "test" test)
+    (section "gpt")
+    (div-textarea "gpt" gpt)
+    [:div [:button {:class btn} "create"]]]])
 
 (defn new [request]
-  (t/log! :info "new")
+  (t/log! {:lelvel :info :id (user request)})
   (page
    (problem-form {:db/id -1
                   :problem/valid true
@@ -62,6 +78,26 @@
                   :problem ""
                   :test ""
                   :gpt ""})))
+
+(defn upsert! [params]
+  (ds/put! params)
+  (t/log! {:level :info :data params}))
+
+(comment
+  (ds/qq '[:find ?e
+           :where
+           [?e :problem/valid _]])
+  (:updated (ds/pl 3))
+  :rcf)
+
+(defn create! [{params :params}]
+  (t/log! {:level :info :data params :msg "create!"})
+  (let [params (-> params
+                   (dissoc :__anti-forgery-token "db/id" "problem/valid")
+                   (assoc :db/id -1 :problem/valid true :updated (jt/local-date-time)))]
+    (t/log! {:level :debug :data params :msg "updated params"})
+    (upsert! params)
+    (resp/redirect "/admin/problems")))
 
 (defn problems [request]
   (page