+++ src/hkimjp/konpy2/response.clj
@@ -4,7 +4,7 @@
    [ring.util.response :as resp]
    [taoensso.telemere :as t]))
 
-(def version "0.1.1")
+(def version "0.2.1")
 
 (def ^:private menu "text-xl font-medium text-white px-1 hover:bg-red-500")